Long Live Qcxscheduler.
Qcxscheduler is an implementation of the cxscheduler interface under Qt
environment.

The library core is written by my colleague - Roman Kuskov - for the
intermediate version of the CX protocol around 2006-2007.

This code is given to me (Yury Rogovsky) "as is" in 2008 without any
guarantees, fees or obligations from both sides.

Further development and support of this code will be made by me (I hope)
for the glory of Open Source Projects under the MIT license untill the
opposite is said.